refactor(sales): split deal print component
