+++ frontend/plugins/sales_ui/src/modules/deals/boards/components/common/print/constants.ts
@@ -1,0 +1,10 @@
+import {
+  PAPER_SIZES,
+  PAPER_TYPES,
+} from 'ui-modules/modules/documents/constants';
+
+export const DEAL_DOCUMENT_CONTENT_TYPE = 'sales:deal';
+export const DEALS_LIMIT = 100;
+export const DEFAULT_PAPER_SIZE = PAPER_SIZES.A4;
+export const PRINT_DOCUMENT_LOAD_TIMEOUT = 15_000;
+export const PRINT_PAPER_TYPE = PAPER_TYPES.SHEET;
